CNN_experiment/requirements.txt
torch
torchvision
numpy
matplotlib
tqdm
Pillow
scipy